+++ zarrs/src/array/codec/array_to_bytes/vlen_v2/vlen_v2_macros.rs
@@ -170,6 +170,13 @@ macro_rules! vlen_v2_codec {
                 self as Arc<dyn ArrayToBytesCodecTraits>
             }
 
+            fn decoded_subchunk_grid(
+                &self,
+                _chunk_grid: &zarrs_chunk_grid::ChunkGrid,
+            ) -> Result<Option<zarrs_chunk_grid::ChunkGrid>, zarrs_chunk_grid::ChunkGridCreateError> {
+                Ok(None)
+            }
+
             fn encode<'a>(
                 &self,
                 bytes: ArrayBytes<'a>,