+++ zarrs/src/array/codec/array_to_bytes/zfpy/zfpy_codec.rs
@@ -1,10 +1,11 @@
 use std::num::NonZeroU64;
 use std::sync::Arc;
 
+use zarrs_chunk_grid::ChunkGridCreateError;
 use zarrs_plugin::{PluginCreateError, ZarrVersion};
 
 use super::super::zfp::ZfpCodec;
-use crate::array::{BytesRepresentation, DataType, FillValue};
+use crate::array::{BytesRepresentation, ChunkGrid, DataType, FillValue};
 use zarrs_codec::{
     ArrayBytes, ArrayBytesRaw, ArrayCodecTraits, ArrayToBytesCodecTraits, CodecCreateError,
     CodecError, CodecMetadataOptions, CodecOptions, CodecTraits, PartialDecoderCapability,
@@ -170,6 +171,13 @@ impl ArrayCodecTraits for ZfpyCodecBound {
 impl ArrayToBytesCodecTraits for ZfpyCodecBound {
     fn into_dyn(self: Arc<Self>) -> Arc<dyn ArrayToBytesCodecTraits> {
         self as Arc<dyn ArrayToBytesCodecTraits>
+    }
+
+    fn decoded_subchunk_grid(
+        &self,
+        _decoded_chunk_grid: &ChunkGrid,
+    ) -> Result<Option<ChunkGrid>, ChunkGridCreateError> {
+        Ok(None)
     }
 
     fn encode<'a>(